index.php first pass
Quick & dirty update to home page; will re-read and revise tomorrow.
Couple of points:

1. I'm assuming some sort of basic search functionality will be
implemented before beta (if not I can update this page accordingly)

2. We need to figure out images.  Looks like we're constraining to a
width of 180, which cripples the screenshots I was going to use (as you
can see in the committed images).  So we need to figure those out and
also need to update the carousel toward the bottom.  I'm thinking just
simple shots of the Budgetr page frames in a monitor, tablet, and phone
display